refactor(vpm): move package json related to package_json mod
