+++ vrc-get-vpm/src/package_json/mod.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+pub mod package_json_like;
+#[allow(clippy::module_inception)]
+pub mod package_json;
+pub mod package_manifest;
